add vpn-mode select
