+++ nym-vpn/ui/src/pages/index.ts
@@ -1,5 +1,5 @@
+export * from './home';
 export { default as NavLayout } from './NavLayout';
-export { default as Home } from './Home';
 export { default as Settings } from './Settings';
 export { default as NodeLocation } from './NodeLocation';
 export { default as Error } from './Error';
