+++ nym-vpn/ui/src/types/routes.ts
@@ -1,3 +1,3 @@
-import { routes } from '../constants.ts';
+import { routes } from '../constants';
 
 export type Routes = (typeof routes)[keyof typeof routes];
